refactor: the construction of the output from `error` helper function into it's own class `ValetException`.
Added:

- Added the new `ValetException` class which extends the PHP native `Exception`, and holds various methods to help construct the `Exception` error messages.

Changed:

- Moved the construction of the output from the `error` helper function into the new `ValetException` class as `getError` public method. And refactored some of the code to make it simpler. This is will be the method that `error` helper function will use to obtain the error message.

- Moved `getErrorTypeName` helper function into the new `ValetException` class with `private` visibility.

- Moved the `constructTrace` loop from the `error` helper function into its own private method in `ValetException` class as `constructTrace`. And refactored some of the code to make it simpler.

- Changed `error` helper function to get the constructed `ValetException` error message via `ValetException::getError` method and outputs this to the console.

- Moved the `error` helper function calls from `GithubPackage::handleApiRateLimitExceededError` method into the `try-catch` block of the `GithubPackage::download` method.

- Moved both `handleApiRateLimitExceededError` and `calculateTimeToGithubApiLimitReset` methods from `GithubPackage` class into the new `ValetException` class. So that all error logic handling is in one class.

- Changed the name of `handleApiRateLimitExceededError` to `githubApiRateLimitExceededError` in the `ValetException` class, and added a return array.

- Changed `handleApiRateLimitExceededError` call in the `GithubPackage::download` method to use the new `ValetException::githubApiRateLimitExceededError` method. It uses the array destructuring assignment to get the return array values and assign them to the respective variable.
