+++ cli/Valet/Packages/GithubPackage.php
@@ -99,7 +99,14 @@ abstract class GithubPackage {
 			$responseMsg = $responseBody->message;
 
 			if (str_contains($responseMsg, 'API rate limit exceeded')) {
-				$this->handleApiRateLimitExceededError($responseHeaders, $responseMsg);
+				[$ip, $rateLimit, $timeLeftToReset] = \ValetException::githubApiRateLimitExceededError($responseHeaders, $responseMsg);
+
+				// Print the error messages.
+				error("\n\nThe GitHub API rate limit has been exceeded for your IP address ($ip). The rate limit is $rateLimit requests per hour.\n\n");
+
+				info("\nThe rate limit will reset in $timeLeftToReset.");
+
+				error("API rate limit exceeded", true);
 			}
 			else {
 				$error = "Error Code: $errorCode\n";
@@ -260,67 +267,5 @@ abstract class GithubPackage {
 		if (!empty($matches)) {
 			return $matches[0];
 		}
-	}
-
-	// TODO: Transfer all error handling to a new separate class.
-	/**
-	 * If the GitHub API rate limit has exceeded, we need to handle it.
-	 *
-	 * The API rate limit is 60 requests per hour for unauthenticated requests.
-	 *
-	 * @param array $headers The headers from the response.
-	 * @param string $msg The error message from the response.
-	 */
-	private function handleApiRateLimitExceededError($headers, $msg) {
-		// Get the rate limit.
-		$rateLimit = $headers["X-RateLimit-Limit"][0];
-		// Get the reset time (UTC epoch seconds).
-		$resetTime = $headers["X-RateLimit-Reset"][0];
-
-		$timeLeftToReset = $this->calculateTimeToGithubApiLimitReset($resetTime);
-
-		// Get the IP address from the original error response message using regex.
-		// The regex pattern matches an IPv4 address.
-		// The pattern looks for 1 to 3 digits followed by a dot, repeated 3 times,
-		// and then 1 to 3 digits.
-		preg_match('/\b(?:\d{1,3}\.){3}\d{1,3}\b/', $msg, $matches);
-
-		// Assign the IP address to a variable if the regex matched, otherwise set it to "unknown".
-		$ip = $matches[0] ?: "unknown";
-
-		// Print the error message.
-		error("\n\nThe GitHub API rate limit has been exceeded for your IP address ($ip). The rate limit is $rateLimit requests per hour.\n\n");
-		info("\nThe rate limit will reset in $timeLeftToReset.");
-		error("API rate limit exceeded", true);
-	}
-
-	/**
-	 * Calculate the time left to reset the GitHub API rate limit.
-	 *
-	 * @param string $resetTime The reset time in UTC epoch seconds.
-	 *
-	 * @return string The time left to reset the rate limit in a human-readable format.
-	 */
-	private function calculateTimeToGithubApiLimitReset($resetTime) {
-		// Create new DateTime objects for the reset time and the current time.
-		$reset_time = new \DateTime("@$resetTime");
-		$current_time = new \DateTime("now");
-
-		// Get the difference between the 2 times.
-		$timeDifference = $reset_time->diff($current_time);
-
-		// Get the difference in minutes and seconds.
-		// The DateInterval object has many properties, including minutes and seconds,
-		// which we can directly access.
-		$mins = $timeDifference->i;
-		$secs = $timeDifference->s;
-
-		// Format the minutes and seconds into a human-readable string.
-		// If the minutes or seconds equals 1, we need to use the singular form
-		// of "minute" or "second".
-		$minsTxt = $mins === 1 ? "$mins minute" : "$mins minutes";
-		$secsTxt = $secs === 1 ? "$secs second" : "$secs seconds";
-
-		return "$minsTxt and $secsTxt";
 	}
 }
